feat(week3-task6): add shared folder for spinner, i18n files, utils and contexts
