+++ week-3/project_ts/src/shared/lib/i18n/i18n.ts
@@ -1,8 +1,8 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 
-import enTranslation from "@/locales/en/translation.json";
-import ruTranslation from "@/locales/ru/translation.json";
+import enTranslation from "@/shared/lib/i18n/locales/en/translation.json";
+import ruTranslation from "@/shared/lib/i18n/locales/ru/translation.json";
 
 i18n
   .use(initReactI18next)
