+++ week-3/project_ts/src/shared/lib/utils/local-storage/constants.ts
@@ -1,0 +1,5 @@
+export const STORAGE_KEY = {
+  FAVORITE_BOOKS: 'favoriteBooks',
+  THEME: 'theme',
+  LANG: 'lang'
+} as const;
