+++ week-3/project_ts/src/shared/model/types/contexts.ts
