+++ week-3/project_ts/src/shared/ui/books-container/books-container.module.scss
@@ -1,0 +1,7 @@
+.books-container {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: var(--gap-lg);
+  padding: var(--gap-xl);
+  min-height: 60vh;
+}
