+++ week-3/project_ts/src/shared/ui/books-container/books-container.tsx
@@ -1,0 +1,9 @@
+import { PropsWithChildren } from "react"
+
+import styles from './books-container.module.scss'
+
+export const BooksContainer = ({children}: PropsWithChildren) => {
+  return (
+    <div className={styles['books-container']}>{children}</div>
+  )
+}
