+++ week-3/project_ts/src/shared/ui/favorite-icon/favorite-icon.module.scss
@@ -1,0 +1,4 @@
+.favorite-icon {
+  font-size: var(--text-xl);
+  color: var(--red);
+}
