+++ week-3/project_ts/src/shared/ui/favorite-icon/favorite-icon.tsx
@@ -1,0 +1,9 @@
+import { FaHeart } from 'react-icons/fa'
+
+import styles from './favorite-icon.module.scss'
+
+export const FavoriteIcon = () => {
+  return (
+    <FaHeart className={styles["favorite-icon"]} />
+  )
+}
