+++ week-3/project_ts/src/shared/ui/spinner/spinner.module.scss
@@ -1,0 +1,7 @@
+.spinner-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: var(--gap-xl);
+  min-height: 100px;
+}
