+++ week-3/project_ts/src/shared/ui/spinner/spinner.tsx
@@ -1,9 +1,11 @@
 import { ClipLoader } from 'react-spinners';
 
+import styles from './spinner.module.scss'
+
 export const Spinner = () => (
-  <div className="spinner-container">
-    <ClipLoader 
-      color="var(--blue)" 
+  <div className={styles["spinner-container"]}>
+    <ClipLoader
+      color="var(--blue)"
       size={50}
       speedMultiplier={0.8}
     />
